Load before and after programs, check module count
This adds option handling to receive the before and after program paths, tries
to load them, and then reports whether they have the same count of modules.

This is a fairly silly check, since *.ll and *.bc files are always a single
module, and we're ignoring KLEE's archive (*.a) loading support here. It's just
a place to start our checking.

A very basic sketch of reporting output is included. Most likely this will
become more advanced as more checks are added.