Skip redundant phi values in LVR analysis
Similar to the recent change which skips redundant intrinsics for simple values,
this adds the same skipping for phi nodes. This version is a bit more complex,
as we need to have first visited all blocks in the program so that we're
tracking all ranges, since the phi inputs may come from any other block.