Collect via `dbg.value` intrinsics
Constants were a bit awkward to reach for, and we'll likely want them as `Expr`s
for later calculations. As part of getting access to them and also achieving a
nicer design, this changes KLEE to track metadata values (like constants) as
part of instruction operands. This allows efficiently collecting via `dbg.value`
intrinsics for all kinds of value producers.