Sort ranges in relative program order
This uses the assembly line of the producing instruction or variable intrinsic
to establish a relative ordering for ranges with the same start lines, which
commonly happens for loop induction variables. This ordering is used to produce
generic range IDs to ease matching ranges (since the assembly lines won't match
across program versions).