Track user for each assignment
This adds the "user" to each assignment (the instruction using the producers as
inputs). This will either be a store instruction or a dbg.value intrinsic. This
allows the producer for the store instruction case to be the store's _value
operand_ (as it likely should have been all along). This neatly fixes up a few
spots where we previously had to check things against both the producer and
variable intrinsic to cover both cases, whereas such checks can now just check
the user.