Skip simple function calls in independent mode
This skips some simple function calls in independent mode to improve execution
time and simplify constraints. Functions with pointer arguments or a variable
number of arguments are not yet supported and will assert.

If the skipped function had a return value, it is made symbolic. The calling
function then continues executing past the skipped call.