Skip runtime functions in independent mode
This skips runtime functions (e.g. `memcpy`) in independent mode. The separate
block jumping feature was behaving erratically for such calls, since it may
never return back to the caller once all blocks are visited.

We could alternatively descend into runtime functions and run them normally, but
this can take quite a bit of time in some cases, and it's not yet clear that we
need these more precise values.