Use machine word size for constants and enable truncation
Now that we're printing types for all constants to resolve query parsing issues,
we need to be a bit more explicit with constant sizes. This flips back to using
machine word size as the starting point for constants (as it's easiest to start
wider and then truncate where needed). This also enables truncation for
constants whenever width differs. This appears to be the best approach for
examples we've encountered so far.